requirements.txt
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-MySQL==1.4.0
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.3.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
msgpack==0.6.1
mysql-connector==2.1.6
mysql-connector-python==8.0.15
mysqlclient==1.4.2.post1
protobuf==3.7.0
PyMySQL==0.9.3
PyYAML==3.13
requests==2.21.0
six==1.12.0
SQLAlchemy==1.2.18
urllib3==1.24.1
Werkzeug==0.14.1
